Create DOM content for menu section in different module
The menu tab features different dishes offered by the restaurant.

The Dish class in the module returns the DOM content of each dish
after taking arguments for name, image source, and description.

The returned dish DOM content is appended to a dish list which is
then exported to 'index.js' module.