Add class to nav buttons; add heading; replace div with main element
The nav buttons for home, menu, and about have been given a class
attribute to access them from JS.

The restaurant name has been added as the main heading of the page.

The div with #context id has been changed to a main element as it
seems more fitting for main element to come after header.