Place header content using CSS Grid
The grid has been set up with 3 rows even though we only have two
elements (h1 and nav).
This is done to keep space between the top edge and the h1 heading
by keeping the first row blank and assigning the h1 the second row,
and nav the last row respectively.

This ensures that header is a comfortable distance away from nav
and the top edge while the nav is away from h1 but near the main
element.