Place Menu Tab contents using CSS Grid
The image height and width constraints have been set to 100% so that
they take up the space allotted to them via Grid and no more or no
less than that.

The row allotted to images (second row) has space of 350px in order
to constraint the image size within that space.