Refactor DOM element append order for image & link
Initially, the image and link (<a> tag) were appended separately to
the 'contact' parent div. This, however, gave rise to image and
text link alignment issues.

Therefore, as a fix, the image is now appended to the link or
anchor tag, which is then appended to the parent div.
This eases the task of alignment of the image and text for us.